+++ upload/admin/controller/localisation/country.php
@@ -527,4 +527,39 @@ class Country extends \Opencart\System\Engine\Controller {
 		$this->response->addHeader('Content-Type: application/json');
 		$this->response->setOutput(json_encode($json));
 	}
+
+	public function autocomplete(): void {
+		$json = [];
+
+		// Manufacturers
+		if (isset($this->request->get['filter_name'])) {
+			$this->load->model('localisation/country');
+
+			$filter_data = [
+				'filter_name' => $this->request->get['filter_name'],
+				'start'       => 0,
+				'limit'       => $this->config->get('config_autocomplete_limit')
+			];
+
+			$results = $this->model_localisation_country->getCountries($filter_data);
+
+			foreach ($results as $result) {
+				$json[] = [
+					'country_id' => $result['country_id'],
+					'name'       => strip_tags(html_entity_decode($result['name'], ENT_QUOTES, 'UTF-8'))
+				];
+			}
+		}
+
+		$sort_order = [];
+
+		foreach ($json as $key => $value) {
+			$sort_order[$key] = $value['name'];
+		}
+
+		array_multisort($sort_order, SORT_ASC, $json);
+
+		$this->response->addHeader('Content-Type: application/json');
+		$this->response->setOutput(json_encode($json));
+	}
 }