+++ CHANGELOG.md
@@ -21,7 +21,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0).
 - `admin` Lightweight library `Jsvectormap` for Dashboard map visualization.
 - `admin` Highlighting active filter buttons.
 - `admin` Added top pagination.
-- `admin` Added IP filter autocomplete for online report.
+- `admin` Added `IP` filter autocomplete for online report.
+- `admin` Added `Country Name` filter autocomplete for Countries list.
 - `catalog` [#15479](https://github.com/opencart/opencart/pull/15479) - Support multi-word product name search.
 - `catalog` Button to change an order on the checkout page.